ai-service/requirements.txt
Django==4.2.26
djangorestframework==3.14.0
django-cors-headers==4.3.0
Pillow==10.3.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
scikit-learn==1.3.0
python-dotenv==1.0.0
gunicorn==22.0.0
